Keep displayed movies list on the MainActivity's recyclerview when the screen is rotated.
Before this change, the movies already displayed before the screen rotation disappeared because
the adapter object, that held them, was destroyed and recreated after this rotation.
Besides, the first load of movies has been moved from the MainActivity to the MainViewModel
to prevent getting new movies when the screen is rotated and this activity is
recreated.
Also, the movies filters logic has been moved from the MainActivity to the MainViewModel.