Add an offline first architecture for the main list of movies.
Create a local storage for movies using a local database with Room and DataStore.
The movies coming from the database are now the source of truth for the main list of movies.
Also add the appropriate repositories and datasources to manage the different sources of data: remote API and local storage.
In addition, the code of the application has been reorganised and some names of variables, methods and classes have been renamed to adapt them to the offline first architecture.