Change the TheMovieDB API authentication type.
The previous authentication type was based on appending the API key as a query parameter in the URL of a request using an OkHttp interceptor.
Now the authentication type consists of adding the Authorization header with a "Bearer token" as its value, using an OkHttp "interceptor". This token is called "Access Token Auth" in the TheMovieDB API.